Rename processor file to makeProcessor.
